move the initial_condition_converter to the converters module
